Fix g++ warning msg
git-svn-id: svn://irodssvn.ucsd.edu/trunk@3936 ea137bd6-aa86-47b6-b90d-b101d9f19a90

+++ iRODS/server/core/src/xmsgLib.c
@@ -372,7 +372,7 @@ irodsXmsg_t **outIrodsXmsg)
 
     if (msgNumber != ANY_MSG_NUMBER) {
         while (tmpIrodsXmsg != NULL) {
-	    if (tmpIrodsXmsg->sendXmsgInfo->msgNumber == msgNumber) break;
+	    if ((int) tmpIrodsXmsg->sendXmsgInfo->msgNumber == msgNumber) break;
 	    tmpIrodsXmsg = tmpIrodsXmsg->tnext;
 	}
     }
